+++ PacientService/Entities/Person.cs
@@ -34,7 +34,7 @@ namespace PacientService.Entities
         public DateTime DateChangePerson { get; set; } = DateTime.Now;
 
         [Display(Name = "ID в промеде")]
-        public int? IdPromedPerson { get; set; } = 0;
+        public Int64? IdPromedPerson { get; set; } = 0;
 
         [Display(Name = "ID пол")]
         public int? Sex_idPerson { get; set; } = 0;
